separate the statistics part from the table module
