+++ src/lib.rs
@@ -4,6 +4,7 @@ pub mod csv;
 mod datatypes;
 mod memory;
 pub mod record;
+mod stats;
 mod table;
 pub(crate) mod util;
 
@@ -12,7 +13,8 @@ pub use datatypes::{
     DataType, Field, Float64Type, Int32Type, Int64Type, Schema, TimeUnit, TimestampSecondType,
     UInt32Type, UInt8Type,
 };
+pub use stats::{ColumnStatistics, Description, Element, ElementCount, FloatRange, NLargestCount};
 pub use table::{
-    BinaryArrayType, Column, ColumnType, Description, Element, Float64ArrayType, Int32ArrayType,
-    Int64ArrayType, Table, UInt32ArrayType, UInt8ArrayType, Utf8ArrayType, DEFAULT_NUM_OF_TOP_N,
+    BinaryArrayType, Column, ColumnType, Float64ArrayType, Int32ArrayType, Int64ArrayType, Table,
+    UInt32ArrayType, UInt8ArrayType, Utf8ArrayType,
 };
